django-rest/requirements.txt
djangorestframework==3.14.0
markdown==3.4.1
django-filter==22.1
django-cors-headers==3.13.0
django-rest-swagger==2.2.0
psycopg2-binary==2.9.5